Moved header files, added Windows MinGW build support, updated codestyle & 2019 becomes 2020
