+++ src/color.cpp
@@ -1,6 +1,6 @@
 /**
 * MIT License
-* Copyright (c) 2019 Matter Team
+* Copyright (c) 2020 Matter Team
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,7 @@
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#include "../include/color.h"
+#include "color.h"
 
 // Color
