+++ src/ui/layout/box.h
@@ -1,6 +1,6 @@
 /**
  * MIT License
- * Copyright (c) 2019 Matter Team
+ * Copyright (c) 2020 Matter Team
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,16 +20,43 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#ifndef UI_PUSHBUTTON_H
-#define UI_PUSHBUTTON_H
+#ifndef UI_BOX_H
+#define UI_BOX_H
 
-#include "../units.h"
-#include "label.h"
+#include "../control.h"
+#include "../../units.h"
+#include <math.h>
+#include <vector>
 
-class PushButton: public Label
+enum class BoxAxis {
+    Horizontal,
+    Vertical
+};
+
+class Box: public Control
 {
 public:
-    PushButton();
+    Box(BoxAxis axis);
+
+    BoxAxis getAxis();
+    void setAxis(BoxAxis axis);
+
+    void addChildControl(Control *child);
+
+    virtual void render(SDL_Renderer *context);
+    virtual void processEvents(SDL_Event* event);
+    virtual void windowSizeChanged();
+    virtual void stateChanged();
+
+private:
+    typedef Control super;
+
+    std::vector<Control*> m_childControls;
+
+    int m_childSpacing = 0;
+    BoxAxis m_axis = BoxAxis::Horizontal;
+
+    void m_updateChildrenFrame();
 };
 
 #endif
