+++ src/ui/window.h
@@ -1,6 +1,6 @@
 /**
  * MIT License
- * Copyright (c) 2019 Matter Team
+ * Copyright (c) 2020 Matter Team
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
